connlib: rename module event_queue to bounded_queue
